requirements.txt
certifi==2022.12.7
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
idna==3.4
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
openai==0.8.0
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.27.1
six==1.16.0
tqdm==4.64.1
urllib3==1.26.14
zipp==3.6.0
